requirements.txt
flask
flask-sqlalchemy
python-dotenv
openai-whisper
transformers
torch